GH-1333 merger WalletProxyContractProxy and GnosisSafeContractProxy
